Add Development Container configuration
Devcontainers are a standard[0] supported by Visual Studio Code[1] and
GitHub[2] which allows to ship full development environments as
containers.

After this change, VS Code is going to prompt if it should run the whole
project in container. Then the whole environment and terminals are
running inside container.

Thanks to devcontainers, we can enforce versions of dependencies for all
developers, including:

* Rust, Cargo
* Anchor
* NodeJS, Typescript
* Solana validator

[0] https://containers.dev/implementors/json_reference/
[1] https://code.visualstudio.com/docs/devcontainers/containers
[2] https://github.com/devcontainers