tests: Run without Docker (#186)
* tests: Run without Docker

Since we have a reliable devenv which doesn't require containerization,
stop using docker in `yarn *test*` commands.

This change doesn't impact devcontainer. It's affects only running
tests on host systems, where Docker is not required anymore.

* runTest/NoAccounts.sh - changed spl noop path to  test-env

* scripts: Switch from bash to sh

Bash is deprecated in macOS. Some implementation of Bourne shell is
going to be present on all of the systems, so let's switch to it.

* scripts: Add support for macOS on Intel

Previously we assumed that x86_64 architecture means Linux. But
that's not the case. This change covers the Linux/macOS and
x86_64/arm64 support matrix entirely.

---------

Co-authored-by: ananas-block <flow23@protonmail.com>