feat(macros): Add aligned-sized crate with `aligned_sized` macro
The purpose of this macro is to wrap Rust structs, calculate the
size of all its fields and implement a `LEN` constant holding
the size.

Thanks to it, we can use those `LEN` constants as parameters to
`space` keys in Anchor instructions without having to calculate
the sizes manually!

As easy as doing:

```rust
pub MyAccount {
    pub foo: u64,
    pub bar: u64,
}

pub struct MyInstruction<'info> {
    #[account(
        init,
        seeds = [b"foo"],
        bump,
        space = MyAccount::LEN,
        payer = signer,
    )]
    pub my_account: Account<'info, MyAccount>,
}
```

There is nothing related to alignment yet, but we are going to
follow up with automatic creation of padding fields in this macro
as well, therefore the name of the crate mentions alignment.