fix(verifier-sdk): Truncate the integrity hash instead of applying modulus
Previously, we were applying modulus on the SHA-256 integrity hash
to fit it in the `Fr` modulo space, making it usable in our
arithmetic circuits.

However, such approach introduces a very small, but stil existing,
chance for a collision (2 unrelated hashes producing the same
value after applying modulus).

Here we switch to a bit safer approach of truncating the hash to
31 bytes (which is guaranteed to fit in `Fr` modulus), reducing
a probability of collision.