feat(zk.js): Fetch PSP verifier program ID in UTXO from the IDL
Previously we were using the verifier program lookup tabke to fetch
the verifier program IDs, but that lookup table didn't include PSPs.

For now, we are not considering building any PSP registry or lookup
table, so let's remove that lookup table for now and use the app
IDL to get the ID of the PSP's verifier.