feat: add base tsconfig and implement its usage in the workspace. (#313)
* Add base tsconfig and implement its usage in the workspace.

This commit includes the creation of a base tsconfig.json file in a new shared directory (tsconfig/**). The purpose of this configuration is to standardize the TypeScript compiler options across different parts of our codebase. This base configuration has been added as a dependency in the prover.js and updated in the pnpm-lock.yaml and pnpm-workspace.yaml files. It will ensure consistent TypeScript compilation settings within all projects, promoting uniformity and reducing potential compilation issues.

* Inherit prover & zk.js from base tsconfig.json + make tsconfig more strict.

* formatting