Refactor test files for clearer error assertion
This commit removes 'nukeDB.ts' and 'setup.jsdom.ts' from the 'include' section of tsconfig.json and relocates them to the 'src' folder. This relocation is likely meant to align with the standard structure of Node.js projects and does not impact the test running process.

Additionally, assertions in 'functional_test.ts' have been refactor to check if 'error' is not false before calling 'message.includes', providing extra layer of error-checking and ensuring that there will not be a Type Error if 'res.error' is undefined.