Correct UTXO appDataHash assignment in zk.js
Modified the assignment of this.appDataHash in the unspent transaction output (UTXO) class of zk.js source file. The change involves calling the wasmPoseidon hash function directly and changing the endian configuration from "le" to "be". This change is done for correctness, as the previous version was introducing computation errors due to the wrong endianess specified and the unnecessary stringifyBigInts operation.