feat: Atomic transactions (leveraging the Poseidon syscall) (#367)
* chore: Move light-merkle-tree to the monorepo

This way it's going to be easier to make changes which affect
other crates.

* style(merkle-tree): Remove the "solana" feature

We are not going to use the Merkle tree outside Solana context,
so keeping this feature only bloats the code.

* feat: Atomic transactions (leveraging the Poseidon syscall)

tl;dr: One Solana transaction performing one Light Protocol
transaction, together with the Merkle tree update.

Thanks to Poseidon syscall in Solana 1.17, we can perform atomic
transactions without necessity to update the Merkle tree in a
separate step.

Make it an optional feature, which can be enabled by:

Building the on-chain program with atomic-transactions feature.
E.g. anchor build -- --features atomic-transactions.
Enabling atomicTransactions flag in the User class in zk.js.