chore: Reorganize the code related to Merkle tree implementation
Before this change, the `merkle-tree` directory contained a single
crate with an append-only sparse Merkle tree implementation.

We are going to introduce more Merkle tree implementations in the
future, like concurrent Merkle tree and indexed Merkle tree. These
new implementations are going to use two elements of the sparse MT
implementation:

* Precomputed zero bytes.
* `Hasher` trait, which allows to easily switch the hash function
  in the MT.

Theferore, in this change, we move these to a separate crate,
called `light-hasher`.

Afterwards, the `merkle-tree` directory contains the following
subdirectories:

* `hasher` (containing `light-hasher` crate)
* `sparse` (containing `light-sparse-merkle-tree` crate)