feat: Add keccak hasher
Although it's not going to be used directly in Light Protocol (at
least for now), it will be useful for testing and comparing whether
our Merkle tree implementations are in line with other implementations
like spl-concurrent-merkle-tree (which uses keccak) or Ethereum rollups
(tending to use keccak). Our Merkle tree is agnostic towards hashing
algorithms, but the other implementations we mentioned are not.

By "are in line" we mean whether they work the same way and produce
the same roots for the same operations.