feat: Add concurrent Merkle tree
Concurrent Merkle tree is the concept of updateable MT which is
described in the whitepaper[0] written by gigabrains from Solana
Labs and Metaplex.

It allows for concurrent updates (of already inserted leaves) of
the tree without any locks and queues.

Incorporating it into Light Protocol is a huge upgrade from our
previous append-only sparse Merkle Tree implementations.

Although we don't need updates for transactions (append-only is
sufficient there), we are going to need updates when imlementing
indexed Merkle tree for nullifiers[1]. This feature is a preparatory
change for implementing indexed Merkle tree.

[0] https://drive.google.com/file/d/1BOpa5OFmara50fTvL0VIVYjtg-qzHCVc/view
[1] https://docs.aztec.network/concepts/advanced/data_structures/indexed_merkle_tree