ci: Run Rust liners both on stable and nightly toolchains
Running clippy nightly is often necessary on macOS (thiserror[0]
fails to build on Rust stable on macOS, we need to investigate why,
it's not in the scope of this change).

But well, new features of clippy nightly are also pointing out some
good coding practices, so we can also enable it anyway and just keep it.

[0] https://crates.io/crates/thiserror