feat: Add psp-account-compression program with address space
Add a new psp-account-compression program which is going to contain all
instructions related to generalized account compression.

As the first functionality, add address space Merkle tree - an indexed
Merkle tree with corresponding program instructions, which stores
address spaces (as ranges). It works the same way that we envisioned
nullifier Merkle tree to work.