perf(merkle-tree): Make no distinction between first and other appends
It's not necessary. Even though spl-concurrent-merkle-tree is doing that,
it's really not. The only thing needed for the usual append algorithm to
handle the first insertion gracefully is to use 0 as the node index
in hash calculations. This is more efficient and less confusing than
treating it as an update.