ci: Pin Rust nightly to 2024-02-01
The newest Rust nightly removed the `stdsimd` feature in favor of
more grained features (rust-lang/rust#117372).

However, the version of `ahash` we depend on, still requires that
feature. New release of `ahash` (0.8.7) contains a fix, but we can't
upgrade to it without upgrading Anchor.

Pinning Rust nightly seems to be the less intrusive solution for now.