chore: Rename `update_proof_or_leaf` to `update_proof` (#460)
Because that's the only thing this function does - it updates the
proof based on the new changelogs.

The confusion came from similarly named `update_proof_or_leaf`
function in the SPL implementation and whitepaper, which returns a leaf
when patching the changelog is impossible (because a new changelog entry
affects the leaf we want to update). But we are not doing it - we found
it confusing and we decided to handle that edge case by returning
an error.