perf: Reduce number of hash operations in batch append
The optimization is about finding a "fillup index" for each
non-terminal leaf in the batch. The fillup index is the node index,
until which we compute the Merkle path. All the nodes of that
path, which are above the fillup index, are not being computed.

The fillup index is determined by checking trailing ones in
the bit representation of the given leaf's index.

The full Merkle path is being computed only for the last leaf
of the batch. Then for each previous Merkle path, we copy the
missing nodes.

This works, because not filling up nodes above the fillup index
doesn't affect the correctnes of computing next Merkle paths.
The nodes which we don't fill up for the given leaf, would be
overwritten by one of the next appends anyway.

What we gain here is less hashing operations we have to perform.

The only theoretical downside is that all Merkle paths
(changelog entries), except the last one, are not exactly correct.
But that doesn't really matter, because as long as any off-chain
consumer (client, relayer) updates a non-sparse Merkle tree using
the whole batch of changelog entries, the final state of the
Merkle tree is going to be correct. Every next Merkle path is
going to correct every incorrect node from the previous
iteration.