perf: Use a generic index type in `IndexingArray` and `IndexingElement`
Depending on context, we prefer either `u16` or `usize`/`u64` there.

`u16` is absolutely fine for on-chain queues (indexing arrays),
because the current limit is 2800. Using a smaller type than `usize`
saves us some space.

However, the same cannot be done for off-chain indexing arrays or
Merkle trees, where the number of elements equals to number of
leaves. Currently, the height of the Merkle tree is 22 which gives
over 4 million leaves.