feat: Add psp-account-compression program with address space (#463)
* feat: Add psp-account-compression program with address space

Add a new psp-account-compression program which is going to contain all
instructions related to generalized account compression.

As the first functionality, add address space Merkle tree - an indexed
Merkle tree with corresponding program instructions, which stores
address spaces (as ranges). It works the same way that we envisioned
nullifier Merkle tree to work.

* perf(utils): Don't reverse the whole array in `bigint_to_be_bytes`

Just fill the array with bytes in correct order once.

* style: Remove leftover logs

* test: Improve error handling for indexed MT updates

* Update relayer's MT only if the update of on-chain MT was
  successful.
* Accumulate all errors inside relayer.

* test: Add more test cases for address space MT

* perf: Use a generic index type in `IndexingArray` and `IndexingElement`

Depending on context, we prefer either `u16` or `usize`/`u64` there.

`u16` is absolutely fine for on-chain queues (indexing arrays),
because the current limit is 2800. Using a smaller type than `usize`
saves us some space.

However, the same cannot be done for off-chain indexing arrays or
Merkle trees, where the number of elements equals to number of
leaves. Currently, the height of the Merkle tree is 22 which gives
over 4 million leaves.

* fix: Correct the Poseidon hasher error message

* test(utils): Add more tests

* chore(utils): Move `BigInt`-related functions to `bigint` module

* fix(programs): Rename address -> account_compression

* fix: Update imports from `light_utils`