feat: Make Merkle tree size defined at runtime
Allow the following parameters of Merkle trees to be defined at runtime:

* height
* changelog size
* roots size

They can be defined dynamically in the code, but only once and only before
convertion (from raw account bytes) or initialization!

This way, we allow to resize the Merkle tree accounts at our will and
we get away with necessity of using const generics to define the sizes.

The whole convertion operation is unsafe, ensuring that the account
size is correct is our responsibility.