chore: Update Solana to 1.18.11
We do it mainly for the new definition of `TransactionSimulationDetails`
in solana-banks-interface 1.18, which contains `inner_instructions`[0].
It will allow us to mock the indexer in Rust tests properly - by fetching
noop program instruction data. This way we could stop returning the
events in instructions and use only the noop program.

[0] https://docs.rs/solana-banks-interface/1.18.11/solana_banks_interface/struct.TransactionSimulationDetails.html