refactor: update nullify_leaves instruction to emit nullify events (#716)
* refactor: update nullify_leaves.rs to emit nullify events

This commit refactors the `nullify_leaves.rs` file to emit nullify events when inserting nullifiers. It introduces a new struct `Nullifier` and an enum `NullifyEvent` to represent the nullify events. The `insert_nullifier` function now creates `NullifyEvent` objects for each inserted nullifier and adds them to a vector. The vector of nullify events is then emitted using the `emit_indexer_event` function.

* update IDLs

* add unit test and renamed Nullifiers to NullifyEvents

---------

Co-authored-by: ananas-block <jorrit@lightprotocol.com>