fix: Handle conflicting changelogs in concurrent Merkle tree (#699)
* fix: Handle conflicting changelogs in concurrent Merkle tree

Restore the usage of `ConcurrentMerkleTreeError::CannotUpdateLeaf` error.
Throwing an error when any changelog entry that modifies the same leaf we
try to update is a right thing to do.

In case of indexed Merkle tree, we can ignore this error. We are already
patching the leaf using the indexed changelog.

Other changes:

* Fix `test_100_nullify_mt`:
  * Each Merkle tree update was pointing to old `changelog_index`. And
    the crank Merkle tree was not being updated after each operation.
* Implement `Clone` in `BoundedVec`, to make cloning of Merkle proofs
  possible.
* Add more tests:
  * Check whether elements returned in `ConcurrentMerkleTree::changelog_entries`
    are as expected.
  * Test proving that we need to `skip(1)` during Merkle proof update.
  * Test with an update on wrapped-around changelog.

* test: Assert a successful update with wrapped-around changelog

The fact that changelog was wrapped around should not affect
correctness of updates when the changelog index and proof are
updated.

* test: Add `get_path_of_leaf` method in `reference::MerkleTree`

We already had `get_proof_of_leaf` for Merkle proofs. The new method
does work similarly, but returns Merkle paths.

* chore: Remove the `light_bounded_vec::Pod` trait

Our vector implementations can store any type. Having the `Pod`
trait doesn't have sense anymore.

* test: Assert full changelog entries in `changelog_entries_random`

* chore: Make the description of `ChangelogEntry::index` more clear