fix: Utilize all value cells of hash set
Before this change, we were never able to fill the hash set to the full
capacity. The `HashSetError::Full` was being returned before the value
array was getting actually full.

That's because `self.capacity_values` was used the modulus for quadratic
probes. Because of this mistake, the additional indices from `self.indices`,
higher than the capacity of `self.values`, were never used. Therefore,
load factors had no effect.

This change fixes it by:

* Using `self.capacity_cells` as modulus.
* Moving the quadratic probe logic to the new `probe_index` method, to
  avoid duplicating the code and having a risk of such bugs in multiple
  places.