refactor: Remove the indices array from hash set
The idea behind indices array was to save space by including additional
buckets for load factor only with small integer types. But it was
generating other problems:

* Picking a value bucket to reuse (due to its expiration) often meant
  that two (or more) index buckets would point to the same value. The
  more often that happens, the less efficient quadratic probing becomes.
* Managing the indices array is an additional effor and might be
  error-prone, especially because no "tratidional" hash set reference
  implementation does that, it was purely our invention.

There were also actual bugs in the mechanism - which are going to be gone
by removing the indices array:

* `capacity_values` (instead of `capacity_indices`) was often used as a
  modulus for quadratic probes.
* The rotating `next_value_index` was still used even in the full hash
  sets in some cases, which led to weird situations.

Therefore, remove the indices array and just apply the load factor to the
main hash set bucket array directly.

Overall, this pull request makes the utilization of hash set more
efficient than it was before. Number of elements which can be inserted
to the set with capacity 6857 stays around 5000.

Before we could insert around 3600 elements to a hash set with indices
capacity 6857 and value capacity 4800.

Other changes:

* Make `unsafe` code self-contained by moving the pointer arithmetics
  to `get_bucket` and `get_bucket_mut` methods.
* Provide `HashSetCell::is_valid` method which checks the sequence
  number. This makes the sequence number checks less error-prone, because
  the comparison is done only in that one place, the other methods are
  supposed to use `is_valid` instead of doing that on their own.

Co-authored-by: ananas-block <jorrit@lightprotocol.com>