feat: Add `append_with_proof` and `apend_batch_with_proofs` methods to CMT
These methods allow for optional return of Merkle proofs.

Our concurrent Merkle tree stores only Merkle paths (in changelog), but
not proofs. Proofs are not needed for the CMT itself, but are needed
for indexed Merkle tree concurrency.

To avoid re-implementing the whole `append` operation, we provide the
`*_with_proof*` method variants, where the caller is supposed to pass
an initialized proof vector or a slice of initialized proof vectors.