fix: Add an explicit `BoundedVec::clone_from` implementation
The default implementation of `clone_from` was causing memory corruption.
Fix that by providing an explicit one which handles capacities correctly.

In case capacities are equal, it just copies the data from one vector to
another.

In case they are not, it starts with reallocating the mutated vector
before copying the data.