feat: Address Merkle tree concurrency (#824)
* feat: Add `append_with_proof` and `apend_batch_with_proofs` methods to CMT

These methods allow for optional return of Merkle proofs.

Our concurrent Merkle tree stores only Merkle paths (in changelog), but
not proofs. Proofs are not needed for the CMT itself, but are needed
for indexed Merkle tree concurrency.

To avoid re-implementing the whole `append` operation, we provide the
`*_with_proof*` method variants, where the caller is supposed to pass
an initialized proof vector or a slice of initialized proof vectors.

* fix: Add an explicit `BoundedVec::clone_from` implementation

The default implementation of `clone_from` was causing memory corruption.
Fix that by providing an explicit one which handles capacities correctly.

In case capacities are equal, it just copies the data from one vector to
another.

In case they are not, it starts with reallocating the mutated vector
before copying the data.

* feat: Address Merkle tree concurrency

Introduce concurrency to indexed Merkle tree by storing an "indexed
changelog entry" structure, which contains:

* The affected element.
* Merkle proof.
* Changelog index (in concurrent MT).

Use indexed changelog to update low elements. In case low element needs
to be updated, swap the element, proof and changelog index.

Swapping of changelog index is needed for the main CMT changelog to
work correctly.

* feat(programs): Introduce `indexed_changelog_index`

To leverage the address Merkle tree concurrency, programs need to take
and pass not only `changelog_index`, but also `indexed_changelog_index`.

* feat(forester): Introduce `indexed_changelog_index`

To leverage the address Merkle tree concurrency, forester needs to take
and pass not only `changelog_index`, but also `indexed_changelog_index`.

* chore: Regenerate IDL

* test: Specify `indexed_changelog_index`

It was introduces in programs, needs to be used in tests as well.

* test: Make `perform_change_log_test` more generic

Take a slice of addresses as arguments, so it's possible to use it to
test more than 2 actors. Perform the updates in a loop.

* fix: Handle double-spend and new low elements in `patch_elements_and_proof`

There are 3 cases which the new elements, introduced in changelog entries,
can fall in:

Already handled:

* An element higher than the element we want to insert. We were handling
  that case already.

Handled by this change:

* An element equal to the element we want to insert. This case means that
  we are trying to double-spend.
* An element lower than the element we want to insert. In this case, it
  becomes the low element. The whole patching procedure needs to be
  started from scratch with that new low element.

* test: Add more details to comments.

* test: Indexed Merkle tree update with 3 actors

* fix: Don't handle double-spend during proof patching process

It's fine to let the Merkle tree detect that afterwards. By not doing it
too early we perform less comparisons.

* fix: Always update the `new_element.index` using `self.next_index()`

Index provided by the user might be outdated. For safety, let's just
always update it with `next_index()` value.

* fix: Don't remove changelog entries

It just doesn't make sense and breaks the concurrency if there are more
than 2 parties wanting to perform update with the same
`indexed_changelog_index` (having equally outdated proofs).

* chore: Remove leftover `println`s

* test: Randomized test for indexed Merkle tree concurrency

* doc: Make comments in CMT about Merkle proofs complete

* chore: Fix build of test-utils

* perf: Don't search for new low element changelog entry

When we find a new low element, we already now that the next changelog
entry (+1) corresponds to its creation.

* chore: Regenerate IDL

* regenerated idl

* chore: Regenerate address MT account

* fix: `invalid_merkle_tree` test

Pass the `indexed_changelog_index`, so the forester util doesn't
try to fetch the account.

* test: Define `indexed_changelog_index` in registry test

* fix buidl

* chore: Update IDL

* test: Pass `indexed_changelog_index`

---------

Co-authored-by: ananas-block <jorrit@lightprotocol.com>
Co-authored-by: Swen <swen.schaeferjohann@code.berlin>