Update README.md and add key in .env.example
The changes in the README.md file expand the Forester service's scope to include dealing with Address Merkle Trees, and ways of setting key in the environment. The update also includes new commands for the 'Usage' section to manually nullify address merkle tree leaves. In addition, a new .env.example file has been added to provide a template for environment variables.