fix: Don't use `fillup_index`
`fillup_index` used to be our arbitrary limit of hashes computed for
each non-terminal Merkle path in a batched update. However, it didn't
work reliably.

`fillup_index` limited to `next_index().trailing_ones()` was resulting
in `filled_subtrees` not being updated in case when `fillup_index` was 0.

Incrementing it by 1, which we used to do, was a workaround which worked
for the most cases in a sense of producing correct roots, but the
non-terminal Merkle paths were incorrect - they contained one node too
many.

Here we are introducing a different approach where we always iterate in
`0..self.height` range, but we break the loop as soon as we encounter a
node on the left. Before breaking the loop, we update `filled_subtrees`.
Therefore, both `filled_subtrees` and all Merkle paths are correct.