fix(macros): Don't include original type size when `size` attribute is used (#1108)
Before this change, `#[aligned_sized]` macro was still including the size
of the original type even on fields using the `size` attribute.

That was an incorrect behavior. The intention of the `size` attribute is
usage for vectors, or other heap-backed types, which we want to serialize
in Solana accounts. Therefore, providing an arbitrary size is necessary.