refactor: Add more common types to `light-sdk`
Add more types, which are going to be used by third-party applications,
which were until now defined in program crates:

- `address`
  - `NewAddressParams`
  - `NewAddressParamsPacked`
- `compressed_account`
  - `CompressedAccount`
  - `CompressedAccountData`
  - `CompressedAccountWithMerkleContext`
  - `PackedCompressedAccountWithMerkleContext`
  - `OutputCompressedAccountWithPackedContext`
- `event`
  - `MerkleTreeSequenceNumber`
  - `PublicTransactionEvent`
- `merkle_context`
  - `QueueIndex`
  - `MerkleContext`
  - `PackedMerkleContext`
- `proof`
  - `CompressedProof`
  - `ProofRpcResult`
- `verify`
  - `CompressedCpiContext`
  - `InstructionDataInvokeCpi`

This is a step in making third-party programs independent from Light
Protocol program crates.

This change does **not** utilize these types in the `verify` function yet.
They're going to be used in follow-up changes, for which we still need
to do some work on RPC and indexer.