refactor: Remove hard dependency on `light-system-program` in SDK
First of all, add more types to the SDK, which are going to be used by
third-party applications, which were until now defined in program
crates:

- `address`
  - `NewAddressParams`
  - `NewAddressParamsPacked`
- `compressed_account`
  - `CompressedAccount`
  - `CompressedAccountData`
  - `CompressedAccountWithMerkleContext`
  - `PackedCompressedAccountWithMerkleContext`
  - `OutputCompressedAccountWithPackedContext`
- `event`
  - `MerkleTreeSequenceNumber`
  - `PublicTransactionEvent`
- `merkle_context`
  - `QueueIndex`
  - `MerkleContext`
  - `PackedMerkleContext`
- `proof`
  - `CompressedProof`
  - `ProofRpcResult`
- `verify`
  - `CompressedCpiContext`
  - `InstructionDataInvokeCpi`

Hide the imports from `light_system_program` behind a feature flag. The
long-term plan is to remove them all together.

Provide a dependency-free, tinier implementation of `TestIndexer` in
the `light-client` crate.