fix: Use `solana-program` in `light-hasher` indiscriminately (#1244)
We were inconsistently using `#[cfg(feature = "solana")]` and
`#[cfg(target_os = "solana")]` in different places, which was leading
to issues when some third-party crate was using `light-hasher` without
`solana` feature, but in sBPF environment.

This feature is a footgun and there is no real benefit of not using
solana-program as a dependency, so just stop using it.

The `solana` feature is stil listed in `Cargo.toml`, but doesn't do
anything. That's because there are still crates depending on that
feature. We can remove it once we are ready to affect them.