feat(sdk): Allow to define instructions without compressed accounts
Make macros aware of a difference between context types:

- `LightContext`, which indicates the usage of compressed accounts.
- `anchor_lang::Context`, which indicates lack of compressed accounts.

By acknowledging that difference, allow developers to have one program
with different instructions, where some of them are using compressed
accounts and some are not.

After this change, the following instruction:

```rust
    pub fn with_compressed_account<'info>(
        ctx: LightContext<'_, '_, '_, 'info, WithCompressedAccount<'info>>,
        name: String,
    ) -> Result<()> {
```

is recognized as the one using cPDAs, because it's using `LightContext`
type.

The following:

```rust
    pub fn without_compressed_account<'info>(
        ctx: Context<'_, '_, '_, 'info, WithoutCompressedAccount<'info>>,
        name: String,
    ) -> Result<()> {
```

is recognized as a regular instruction without cPDAs, because it's using
`Context` type.