chore: csdk macros3 (#2062)
* rent impl for c-PDA.

reuses most of the rent impl of c-token, however we always transfer rent to the program's rent_sponsor pda, even if a different fee_payer paid for rent exemption.

for a rent v1 the benefits of this (saves account space, simpler) outweigh the cons. I think thats fair because the "fee" goes to the program's rent_sponsor PDA, funding the next user's PDA rent exemption.

* added derive_rent_sponsor!(program_id) macro
* allow rust expressions as seeds to add_compressible_instructions!() macro
* allow seed references to other instruction accounts as well as instruction_data
* made macros modular, moved processors to token-sdk and sdk (compress_runtime.rs, decompress_runtime.rs)

added tests for the macros at every level:
* sdk-compressible-test is a manual implementation of the hooks
* csdk-anchor-derived-test uses most smaller macros
* csdk-anchor-full-derived-test uses add_compressible_instructions macro


---------

Co-authored-by: ananas <jorrit@lightprotocol.com>